fix(autocomplete): fix typos and move AutocompleteBase, AutocompleteMultiColumn into
the Autocomplete/components directory
